added confirm function in favorite button and made a logout button and function

+++ scripts/favourites.js
@@ -1,32 +1,40 @@
 
 function addToFavourites(routeId) {
-    const userId = firebase.auth().currentUser.uid; // Get the ID of the currently logged in user
-    
-    let params = new URL(window.location.href); //get the url from the search bar
-    let ID = params.searchParams.get("docID"); 
-    
-    const routeRef = firebase.firestore().collection("routes").doc(routeId); // Get route info
-    const favouritesRef = firebase.firestore().collection("users").doc(userId).collection("favourites").doc(ID); // Create a reference to the user's favourites list
-    favouritesRef.get().then((doc) => {
-        if (doc.exists) {
-            // The product is already in the user's favorites, so remove it.
-            favouritesRef.delete();
-          } else {
-          const routeRef = firebase.firestore().collection("routes").doc(ID);
-          routeRef.get().then((doc) => {
+  const userId = firebase.auth().currentUser.uid; // Get the ID of the currently logged in user
+
+  let params = new URL(window.location.href); //get the url from the search bar
+  let ID = params.searchParams.get("docID");
+
+  const routeRef = firebase.firestore().collection("routes").doc(routeId); // Get route info
+  const favouritesRef = firebase.firestore().collection("users").doc(userId).collection("favourites").doc(ID); // Create a reference to the user's favourites list
+  favouritesRef.get().then((doc) => {
+    if (doc.exists) {
+      // The product is already in the user's favorites, so remove it.
+      const removeConfirm = confirm('This route is already in your favourites. Do you want to remove it?');
+      if (removeConfirm) {
+        favouritesRef.delete();
+        alert('The route has been removed from your favourites.');
+      }
+    } else {
+      const addConfirm = confirm('Do you want to add this route to your favourites?');
+      if (addConfirm) {
+        const routeRef = firebase.firestore().collection("routes").doc(ID);
+        routeRef.get().then((doc) => {
           favouritesRef.set({
-          name: doc.data().name,
-          city: doc.data().city,
-          code: doc.data().code,
-          province: doc.data().province,
-          details: doc.data().details,
-          img: doc.data().img,
-          desc: doc.data().desc
+            name: doc.data().name,
+            city: doc.data().city,
+            code: doc.data().code,
+            province: doc.data().province,
+            details: doc.data().details,
+            img: doc.data().img,
+            desc: doc.data().desc
           });
         });
-        }
-    });
-  }
+        alert('The route has been added to your favourites.');
+      }
+    }
+  });
+}
   
   // Checks if user is logged in and gets user ID to pass to displayFavsDynamically
   firebase.auth().onAuthStateChanged(user => {
@@ -35,7 +43,7 @@ function addToFavourites(routeId) {
       displayFavsDynamically("favourites", userId);
     } else {
       console.log("User is not logged in.");
-    }
+    } 
   });
